server/requirements.txt
django==1.9.7
djangorestframework==3.3.3

markdown
django-filter

